bookly/telegram-bot/requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
prisma==0.13.0